+++ src/com/carolinarollergirls/scoreboard/model/ClockModel.java
@@ -9,6 +9,7 @@ package com.carolinarollergirls.scoreboard.model;
  */
 
 import com.carolinarollergirls.scoreboard.Clock;
+import com.carolinarollergirls.scoreboard.snapshots.ClockSnapshot;
 
 public interface ClockModel extends Clock
 {
@@ -51,24 +52,5 @@ public interface ClockModel extends Clock
 	public void changeMaximumTime(long ms);
 
 	public void setCountDirectionDown(boolean down);
-
-	public class ClockSnapshot {
-		public ClockSnapshot(ClockModel clock) {
-			id = clock.getId();
-			number = clock.getNumber();
-			time = clock.getTime();
-			isRunning = clock.isRunning();
-		}
-		
-		public String getId() { return id; }
-		public int getNumber() { return number; }
-		public long getTime() { return time; }
-		public boolean isRunning() { return isRunning; }
-		
-		protected String id;
-		protected int number;
-		protected long time;
-		protected boolean isRunning;
-	}
 }
